refactor: fold topology controls into adaptive
