+++ Cargo.toml
@@ -5,7 +5,6 @@
 members = [
     "crates/core",
     "crates/adaptive",
-    "crates/adaptive-topology",
     "crates/pii-redaction",
     "crates/cli",
     # Language Bindings
@@ -26,7 +25,6 @@ repository = "https://github.com/NVIDIA/NeMo-Relay"
 [workspace.dependencies]
 nemo-relay = { version = "0.5.0", path = "crates/core", default-features = false }
 nemo-relay-adaptive = { version = "0.5.0", path = "crates/adaptive" }
-nemo-relay-adaptive-topology = { version = "0.5.0", path = "crates/adaptive-topology" }
 nemo-relay-pii-redaction = { version = "0.5.0", path = "crates/pii-redaction" }
 nemo-relay-ffi = { version = "0.5.0", path = "crates/ffi" }
 nemo-relay-cli = { version = "0.5.0", path = "crates/cli" }
