--- crates/adaptive-topology/Cargo.toml
@@ -1,27 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-[package]
-name = "nemo-relay-adaptive-topology"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-description = "Topology, manifold, and adaptive threshold primitives for NeMo Relay."
-readme = "README.md"
-
-[lints]
-workspace = true
-
-[dependencies]
-libm = "0.2"
-serde = { version = "1", features = ["derive"], optional = true }
-
-[features]
-default = ["std"]
-std = []
-alloc = []
-serde = ["dep:serde"]
-
-[dev-dependencies]
-serde_json = "1"
