+++ docs/adaptive-plugin/about.mdx
@@ -54,6 +54,9 @@ If instrumentation is not in place yet, start with
   cache planning accomplishes.
 - [Adaptive Hints](/adaptive-plugin/adaptive-hints) explains request hint injection and how
   downstream model paths can consume the hints.
+- [Topology-Aware Controls Design](/adaptive-plugin/topology-aware-controls-design)
+  documents the internal design, measurable outcomes, and observable signals for
+  convergence, drift, and hint-governor controls.
 
 State, telemetry, tool parallelism, and policy are whole-plugin configuration
 areas. They are documented on [Adaptive Configuration](/adaptive-plugin/configuration) rather
