+++ docs/adaptive-plugin/acg.mdx
@@ -232,7 +232,7 @@ Use `passthrough` when you want ACG observations without provider-specific hint
 translation. Set `provider` to the backend API surface the agent actually calls
 when you are ready to apply cache planning.
 
-ACG convergence uses deterministic topology-inspired stability signals derived
+ACG convergence uses deterministic topology-aware stability signals derived
 from observed prompt structure. It is an adaptive early-stop control for
 learning, not an exact persistent-homology calculation.
