+++ docs/adaptive-plugin/topology-aware-controls-design.mdx
@@ -1,0 +1,83 @@
+---
+title: "Topology-Aware Controls Design"
+sidebar-title: "Topology Design"
+description: ""
+position: 5
+---
+{/* SPDX-FileCopyrightText: Copyright (c) 2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+SPDX-License-Identifier: Apache-2.0 */}
+
+This design records the review contract for the optional topology-aware controls
+inside the Adaptive plugin. The implementation is internal to
+`nemo-relay-adaptive`; it does not add a separate topology crate and it does not
+expose topology primitives as public binding APIs.
+
+## Goals
+
+- Stop ACG learning after repeated stable prompt structure has been observed.
+- Discard stale tool-parallelism plans when the observed tool cohort shape
+  changes sharply.
+- Shed learned adaptive hints below a configurable sensitivity threshold while
+  preserving manual latency-sensitivity overrides.
+- Keep every control disabled by default and observable through existing
+  adaptive state, request metadata, and validation reports.
+
+## Non-Goals
+
+- Exact persistent homology or a general-purpose topology library.
+- New public Rust, Python, Node.js, Go, WebAssembly, or C FFI topology
+  primitives.
+- Changes to NeMo Relay scope semantics, event shape, callback execution, or
+  user callback return values.
+
+## Internal Design
+
+The adaptive crate owns a small internal module, `crate::topology`, with three
+bounded primitives:
+
+- `ConvergenceDetector` tracks a fixed history of Betti-like stability
+  signatures, drift, and error.
+- `DriftDetector` tracks centroid motion for tool cohort feature vectors.
+- `GeometricGovernor` adapts a sensitivity threshold for learned hint
+  injection.
+
+ACG maps each stability analysis result to:
+
+```text
+beta_0 = stable_prefix_length
+beta_1 = total_spans - stable_prefix_length
+drift  = 1 - stable_prefix_length / total_spans
+error  = 1 - average_stability_score
+```
+
+The tool-parallelism learner maps observed tool cohorts to a four-value centroid:
+
+```text
+[cohort_count, unique_tool_count, duplicate_reference_ratio, max_cohort_size]
+```
+
+Adaptive hints use the governor only for learned hints. A manual
+`set_latency_sensitivity()` override still forces hint injection for the current
+request.
+
+## Measurable Outcomes
+
+| Control | Improvement | Observable Signal | Validation |
+|---|---|---|---|
+| ACG convergence | Stable profiles stop consuming new observations after at least `stability_window` epochs instead of always filling the full observation window. | Persisted `StabilityAnalysisResult.converged = true`; later runs reuse cached stability. | `crates/adaptive/tests/integration/topology_convergence_tests.rs` and `crates/adaptive/benches/convergence_bench.rs`. |
+| Tool drift | Stale fan-out plans are rebuilt when cohort drift exceeds `tool_parallelism.drift.threshold`. | The stored `ExecutionPlan` no longer contains old fan-out groups after drift. | `crates/adaptive/tests/unit/tool_parallelism_learner_tests.rs`. |
+| Hint governor | Low-sensitivity learned hints are omitted when below `adaptive_hints.governor.epsilon`. | `nvext.agent_hints` is absent from request headers/body; manual overrides still emit it. | `crates/adaptive/tests/unit/adaptive_hints_intercept_tests.rs`. |
+| Config safety | Invalid thresholds fail before activation. | Plugin validation diagnostics name invalid topology-aware fields. | `crates/adaptive/tests/unit/runtime_tests.rs` and `crates/adaptive/tests/unit/plugin_component_tests.rs`. |
+
+The convergence benchmark prints `observations-to-decision` for a stable
+synthetic prompt profile with and without convergence. The expected property is
+`with_convergence <= without_convergence`; with the default benchmark
+configuration, convergence is eligible after `stability_window = 3` stable
+epochs.
+
+## Rollout
+
+All topology-aware fields default to disabled. A rollout should enable one
+control at a time, validate representative workloads, and use existing adaptive
+state inspection to confirm the observable signals above before enabling the
+next control.
