+++ python/nemo_relay/_native.pyi
@@ -1126,128 +1126,6 @@ class PluginContext:
         for runtime registration callbacks.
     """
 
-class GeometricGovernor:
-    """Adaptive threshold controller using a PD control law.
-
-    Summary:
-        Maintains a sensitivity threshold ``epsilon`` that adapts to the
-        observed deviation rate.
-
-    Description:
-        The governor increases ``epsilon`` when the system wakes too often and
-        decreases it when the system is too sluggish. ``should_trigger`` compares
-        a deviation against the current threshold.
-    """
-
-    def __init__(self) -> None:
-        """Create a governor with the default threshold and gains."""
-        ...
-    @staticmethod
-    def with_epsilon(epsilon: float) -> GeometricGovernor:
-        """Create a governor with a custom initial threshold."""
-        ...
-    def adapt(self, deviation_delta: float, dt: float) -> float:
-        """Adapt epsilon based on the observed deviation over ``dt`` seconds."""
-        ...
-    def should_trigger(self, deviation: float) -> bool:
-        """Return whether ``deviation`` meets or exceeds the threshold."""
-        ...
-    @property
-    def epsilon(self) -> float:
-        """Return the current threshold."""
-        ...
-    @property
-    def adjustment_count(self) -> int:
-        """Return the number of adaptations performed so far."""
-        ...
-    @property
-    def last_error(self) -> float:
-        """Return the error from the most recent adaptation."""
-        ...
-    def reset(self) -> None:
-        """Reset the governor to its initial state."""
-        ...
-    def __repr__(self) -> str:
-        """Return a debug representation of the governor."""
-        ...
-
-class ConvergenceDetector:
-    """Topological convergence detector.
-
-    Summary:
-        Declares convergence when Betti numbers are stable, drift is
-        decreasing, and error falls below a threshold.
-
-    Description:
-        ``record_epoch`` feeds topology metrics from each epoch. ``is_converged``
-        and ``convergence_score`` read the current state.
-    """
-
-    def __init__(self, epsilon: float, stability_window: int) -> None:
-        """Create a detector with the given error threshold and stability window."""
-        ...
-    def record_epoch(self, beta_0: int, beta_1: int, drift: float, error: float) -> None:
-        """Record an epoch's Betti numbers, drift, and error."""
-        ...
-    @property
-    def is_converged(self) -> bool:
-        """Return whether convergence criteria are satisfied."""
-        ...
-    @property
-    def convergence_score(self) -> float:
-        """Return a score in ``[0, 1]`` indicating nearness to convergence."""
-        ...
-    @property
-    def epoch(self) -> int:
-        """Return the number of epochs recorded."""
-        ...
-    @property
-    def last_betti(self) -> tuple[int, int] | None:
-        """Return the most recent ``(beta_0, beta_1)`` pair, if any."""
-        ...
-    @property
-    def last_error(self) -> float | None:
-        """Return the most recent error, if any."""
-        ...
-    def reset(self) -> None:
-        """Reset the detector to its initial state."""
-        ...
-    def __repr__(self) -> str:
-        """Return a debug representation of the detector."""
-        ...
-
-class DriftDetector:
-    """Centroid trajectory drift detector.
-
-    Summary:
-        Tracks the velocity of a 3-dimensional centroid trajectory and reports
-        drift from the predicted position.
-
-    Description:
-        ``update`` records a new centroid and returns the drift from the
-        predicted position. ``is_drifting`` compares the velocity magnitude
-        against a threshold.
-    """
-
-    def __init__(self) -> None:
-        """Create a detector with no previous observation and zero velocity."""
-        ...
-    def update(self, centroid: Sequence[float]) -> float:
-        """Record a centroid and return drift from the predicted position."""
-        ...
-    def is_drifting(self, threshold: float) -> bool:
-        """Return whether the velocity magnitude exceeds ``threshold``."""
-        ...
-    def velocity_magnitude(self) -> float:
-        """Return the magnitude of the current velocity vector."""
-        ...
-    def reset(self) -> None:
-        """Reset the detector to its initial state."""
-        ...
-    def __repr__(self) -> str:
-        """Return a debug representation of the detector."""
-        ...
-
 def create_scope_stack() -> ScopeStack:
     """Create a fresh native scope stack.
